+++ tests/Concerns/WithStrictNullComparisonTest.php
@@ -96,7 +96,7 @@ class WithStrictNullComparisonTest extends TestCase
     /**
      * @test
      */
-    public function exports_empty_cells()
+    public function exports_trailing_empty_cells()
     {
         $export = new class implements FromCollection, WithStrictNullComparison {
             use Exportable;
@@ -130,5 +130,42 @@ class WithStrictNullComparisonTest extends TestCase
         $contents = file_get_contents($file);
         $this->assertStringContains('"a1","","","d1",""', $contents);
         $this->assertStringContains('"a2","","","d2",""', $contents);
+    }
+
+    /**
+     * @test
+     */
+    public function exports_trailing_empty_cells_by_setting_config_strict_null_comparison()
+    {
+        config()->set('excel.exports.strict_null_comparison', false);
+
+        $export = new class implements FromCollection {
+            use Exportable;
+
+            /**
+             * @return Collection
+             */
+            public function collection()
+            {
+                return collect([
+                    ['a1', '', '', 'd1', ''],
+                    ['a2', '', '', 'd2', ''],
+                ]);
+            }
+        };
+
+        $file = __DIR__ . '/../Data/Disks/Local/empty-cells-config.csv';
+
+        $export->store('empty-cells-config.csv');
+
+        $contents = file_get_contents($file);
+        $this->assertStringContains('"a1","","","d1"', $contents);
+
+        config()->set('excel.exports.strict_null_comparison', true);
+
+        $export->store('empty-cells-config.csv');
+
+        $contents = file_get_contents($file);
+        $this->assertStringContains('"a1","","","d1",""', $contents);
     }
 }